Account Module Added

+++ Core/RequestModels/TransactionRequest.cs
@@ -8,9 +8,9 @@ namespace Core.RequestModels
         public int MemberId { get; set; }
         public int ProjectId { get; set; }
         public int InstallmentNo { get; set; }
-        public int DueAmounts { get; set; }
-        public int PayableAmounts { get; set; }
-        public int TransactionAmounts { get; set; }
+        public double DueAmounts { get; set; }
+        public double PayableAmounts { get; set; }
+        public double TransactionAmounts { get; set; }
         public int TransactionType { get; set; }
         public DateTime TransactionDate { get; set; }
         public int VerifiedBy { get; set; }
